Expose versioning info in version plugin
